Add created timestamp to entries and fix failing tests
Tests were failing due to a non-deterministic order in the index.
This was due to the order depending on hashing, which gets the seed
from the environment.

To fix this, the order is now deterministic. First, we sort on the
created timestamp. If that is equal, we sort on the (unique) name.