Remove trash path from path data
There is no reason to include the trash path in the PathData type,
as the information is implicitly included in the index's location
itself. This makes our output nicer (filename only, no redundant
path info) and simplifies the implementation. We do need to take care
to add the path  back when doing existence tests.