Derive Timestamp Eq/Ord
Our first timestamp attempt used ZonedTime, which had no Eq instance,
hence declaring an equivalence class on the serialized string. The
LocalTime we are using does have Eq/Ord instances, so we can just
use those.

We add the time-compat package as it has the (orphan) Hashable
instance for LocalTime.