Swap IO for abstract MonadIO m
Originally implementing everything in terms of bare IO made sense
as the application was very simple and there was no need for abstraction.

While the app is still simple now, we now reach a point where we will want to
abstract the IO, as we will eventually want to test some future print statements
(i.e. partially successful deletes), and this is significantly more difficult if
we have to capture console IO, rather than having a typeclass that stores the logs
in a ref.

As a bonus, this simplifies our Runner logic as we no longer have to manually
pass a text handler.