Simplify index search
Due to the previous change i.e. removing unnecessary checks from the
index search, we can now make our index search total! We are no longer
performing _any_ impure validation on the index because there is
none left!

The only other check was to see if restoring a path would overwrite
an existing one, but now that we are handling exceptions on a per-path
basis, there is no reason to do this ahead of time. Thus we can
defer this check to when we go to actually restore the path.

This dramatically simplifies the code and usage.