Remove TestArgs
The only arg was the test dir, which we can get every time from
the environment. This should be stable within a single test run,
so this probably isn't harmful.

The bonus is we can use it later in other functions (e.g. capture)
without threading the argument through everywhere.