Swap Katip for MonadLogger
Katip brings significant complexity. While this is not a problem in
itself, it also has a limitation that was not appreciated at first:
its rigidity makes it difficult to adapt to different circumstances.

In our case, we want to write golden tests for logs, and timestamps
presents a problem. The obvious solution is to make some typeclass
with 'getCurrentTime :: m Timestamp'. Unfortunately Katip makes this
very difficult as the log formatter is a pure function, whereas the
"push log to dest" function uses IO, so not overloaded.

MonadLogger, on the other hand, gives us a general 'args :: m ()'
function that does everything. This lack of granularity means we have
to do more ourselves, but it makes the above pattern trivial.

We also get the added benefit that our dependency graph / setup
code is simpler.