Change HashSet inputs to new UniqueSeq
For testing and general predictability, it would be nice if we operate
on paths in received order. Using as hashset makes this impossible,
as hashing is non-deterministic. We thus create a new UniqueSeq type
that preserves the original order while maintaining the set's
uniqueness invariant.

Using our new found determinism, we can greatly simplify the functional
tests. We no longer need to use our bespoke "text matcher" type;
instead we can utilize golden tests to do the diffs for us.